+++ src/backend/plugins/tests/organizations/test_name_from_siret_organization_plugin.py
@@ -3,7 +3,7 @@
 import pytest
 import responses
 
-from core.models import Organization
+from core.models import Organization, get_organization_metadata_schema
 from core.plugins.loader import get_organization_plugins
 
 pytestmark = pytest.mark.django_db
@@ -30,6 +30,12 @@ def organization_plugins_settings_fixture(settings):
     get_organization_plugins.cache_clear()
     get_organization_plugins()  # call to populate the cache
 
+    settings.ORGANIZATION_METADATA_SCHEMA = "fr/organization_metadata.json"
+
+    # Reset the model validation cache
+    get_organization_metadata_schema.cache_clear()
+    get_organization_metadata_schema()
+
     yield
 
     # reset get_organization_plugins cache
@@ -37,9 +43,26 @@ def organization_plugins_settings_fixture(settings):
     get_organization_plugins.cache_clear()
     get_organization_plugins()  # call to populate the cache
 
+    settings.ORGANIZATION_METADATA_SCHEMA = None
+
+    # Reset the model validation cache
+    get_organization_metadata_schema.cache_clear()
+    get_organization_metadata_schema()
+
 
 @responses.activate
-def test_organization_plugins_run_after_create(organization_plugins_settings):
+@pytest.mark.parametrize(
+    "nature_juridique,is_commune,is_public_service",
+    [
+        ("123", False, False),
+        ("7210", True, False),
+        ("123", False, True),
+        ("7210", True, True),
+    ],
+)
+def test_organization_plugins_run_after_create(
+    organization_plugins_settings, nature_juridique, is_commune, is_public_service
+):
     """Test the run_after_create method of the organization plugins for nominal case."""
     responses.add(
         responses.GET,
@@ -54,7 +77,11 @@ def test_organization_plugins_run_after_create(organization_plugins_settings):
                             "liste_enseignes": ["AMAZING ORGANIZATION"],
                             "siret": "12345678901234",
                         }
-                    ]
+                    ],
+                    "nature_juridique": nature_juridique,
+                    "complements": {
+                        "est_service_public": is_public_service,
+                    },
                 }
             ],
             "total_results": 1,
@@ -69,10 +96,14 @@ def test_organization_plugins_run_after_create(organization_plugins_settings):
         name="12345678901234", registration_id_list=["12345678901234"]
     )
     assert organization.name == "Amazing Organization"
+    assert organization.metadata["is_commune"] == is_commune
+    assert organization.metadata["is_public_service"] == is_public_service
 
     # Check that the organization has been updated in the database also
     organization.refresh_from_db()
     assert organization.name == "Amazing Organization"
+    assert organization.metadata["is_commune"] == is_commune
+    assert organization.metadata["is_public_service"] == is_public_service
 
 
 @responses.activate
@@ -157,6 +188,10 @@ def test_organization_plugins_run_after_create_no_list_enseignes(
                             "siret": "12345678901234",
                         }
                     ],
+                    "nature_juridique": "123",
+                    "complements": {
+                        "est_service_public": True,
+                    },
                 }
             ],
             "total_results": 1,
@@ -171,7 +206,11 @@ def test_organization_plugins_run_after_create_no_list_enseignes(
         name="12345678901234", registration_id_list=["12345678901234"]
     )
     assert organization.name == "Amazing Organization"
+    assert organization.metadata["is_commune"] is False
+    assert organization.metadata["is_public_service"] is True
 
     # Check that the organization has been updated in the database also
     organization.refresh_from_db()
     assert organization.name == "Amazing Organization"
+    assert organization.metadata["is_commune"] is False
+    assert organization.metadata["is_public_service"] is True